requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
groq
